some component path refactored
