+++ components/ProductSpecs.vue
@@ -21,6 +21,5 @@ export default {
 </script>
 
 <style scoped>
-  @import url('~/assets/css/components/ProductSliderCarouselScoped.css');
-
+  @import url('~/assets/css/components/ProductSpecsScoped.css');
 </style>
